requirements.txt
transformers==4.39.0
torch==2.2.1
tqdm==4.66.2
pandas==2.0.3
scispacy==0.5.4
# Extras
# pipdeptree==2.16.1
# modin==0.23.1.post0

